Ditch CartLineManager
