+++ packages/core/src/Models/Cart.php
@@ -53,18 +53,6 @@ class Cart extends BaseModel
     ];
 
     /**
-     * The "booted" method of the model.
-     *
-     * @return void
-     */
-    protected static function booted()
-    {
-        static::retrieved(function ($cart) {
-            $cart->restoreProperties();
-        });
-    }
-
-    /**
      * Return the cart manager.
      *
      * @return \Lunar\Managers\CartManager
